Update some SHITL and Flightplan code.

+++ src/FlightPlan.hpp
@@ -55,7 +55,7 @@ class FlightPlan{
 
     private:
         float pad_alts[2] = {0.0, 0.0}; //a buffer of past altitudes, the 0th of which will be the pad altitude
-        uint8_t pad_alt_counter = 0;
+        uint32_t pad_alt_counter = 0;
         uint32_t print_timer = 0;
         uint32_t state_timer = 0;
         FlightState state;
